11.23 공부기록 업로드

+++ WinProject/WinProject/Client/struct.h
@@ -5,9 +5,9 @@ struct Vector2
 	float y;
 public:
 	Vector2() : x(0.f), y(0.f){}
-	Vector2(float _x, float _y): x(_x), y(_y){}
-	Vector2(int _x, int _y): x(_x), y(_y){}
-	Vector2(long _x, long _y) : x(_x), y(_y) {}
+	Vector2(float _x, float _y): x((float)_x), y((float)_y){}
+	Vector2(int _x, int _y): x((float)_x), y((float)_y){}
+	Vector2(long _x, long _y) : x((float)_x), y((float)_y) {}
 	Vector2(POINT _pt) :x((float)_pt.x), y((float)_pt.y) {}
 	Vector2& operator=(POINT _pt) {
 		x = (float)_pt.x;
@@ -25,7 +25,7 @@ public:
 	Vector2 operator-(const Vector2& _pt) {
 		return Vector2(x - (float)_pt.x, y - (float)_pt.y);
 	}
-	float Length() { return sqrt(x * x + y * y); }
+	float Length() { return sqrtf(x * x + y * y); }
 	Vector2& Normalize() {
 		float length = Length();
 		assert(length != 0.f);
